refactor(worker-utils): host label + sandbox-id helpers for cross-package reuse
Move the pure sandboxId <-> hostname-label logic (plus sandboxId <-> userId
encoding) from `services/kiloclaw/src/auth/` into
`@kilocode/worker-utils` so `apps/web` can use it to mint per-instance
URLs in PR3 without duplicating the base64url / base32hex encoding.

- New subpath exports: `@kilocode/worker-utils/hostname-label` and
  `@kilocode/worker-utils/sandbox-id`.
- The existing `services/kiloclaw/src/auth/hostname-label.ts` and
  `sandbox-id.ts` become thin re-export shims so the many existing
  `./auth/hostname-label` / `./auth/sandbox-id` imports inside the
  worker don't have to migrate all at once.
- Tests move with the implementation.

No behaviour change.
